# fix(bibigridperf): remove padding when parsing hdparm output